pasted in controllerstech OLED library (ssd1306.c, ssd1306.h, fonts.h)

+++ Core/Inc/spl06-007.h
@@ -101,7 +101,7 @@ uint8_t SPL06_007_Initialise( SPL06_007 *dev, I2C_HandleTypeDef *i2cHandle );
  * DATA ACQUISITION
  */
 uint8_t SPL06_007_calcCompPressure( SPL06_007 *dev );
-uint8_t SPL06_007_calcCompTemp( SPL06_007 *dev, int32_t rawTemp );
+uint32_t SPL06_007_calcCompTemp( SPL06_007 *dev, int32_t rawTemp );
 uint8_t SPL06_007_getRawPressure( SPL06_007 *dev );
 int32_t SPL06_007_getRawTemp( SPL06_007 *dev );
